% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_CDC.R
\name{collect_cdc_batches}
\alias{collect_cdc_batches}
\title{Collect paginated batches from CDC Socrata dataset}
\usage{
collect_cdc_batches(series_id, app_token = NULL, limit = 1000, verbose = FALSE)
}
\arguments{
\item{series_id}{Dataset ID to pull from}

\item{app_token}{Optional API token}

\item{limit}{Page size (default 1000)}

\item{verbose}{Print progress}
}
\value{
A list of tibbles (each page of results)
}
\description{
Collect paginated batches from CDC Socrata dataset
}
\seealso{
\code{get_cdc_batch()}, \code{harmonize_flag_column()}
}
\keyword{internal}
