% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_CDC.R
\name{detect_cdc_structure}
\alias{detect_cdc_structure}
\title{Detect the CDC dataset structure based on column names hash}
\usage{
detect_cdc_structure(df)
}
\arguments{
\item{df}{A tibble returned from CDC data batching.}
}
\value{
A string indicating which cleaning function to apply, or \code{NULL} if no
match is found.
}
\description{
Given a tibble (typically returned from \code{bind_cdc_batches()}), this function
generates an MD5 hash of the column names and uses it to determine which data
cleaning function should be applied.
}
\keyword{internal}
