% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_CDC.R
\name{get_cdc_batch}
\alias{get_cdc_batch}
\title{Get a single batch (page) of CDC Socrata data}
\usage{
get_cdc_batch(series_id, offset = 0, limit = 1000, app_token = NULL)
}
\arguments{
\item{series_id}{Socrata dataset ID (e.g., "489q-934x")}

\item{offset}{Row offset to begin from (default is 0)}

\item{limit}{Number of rows per page (default is 1000)}

\item{app_token}{Optional Socrata API token}
}
\value{
A tibble containing the batch of rows
}
\description{
Get a single batch (page) of CDC Socrata data
}
\seealso{
\code{collect_cdc_batches()}
}
\keyword{internal}
